requirements.txt
APScheduler==3.10.3
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.2
Flask-APScheduler==1.12.4
gevent==23.7.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tzlocal==5.0.1
urllib3==2.0.4
werkzeug==2.3.7
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
